requirements.txt
numpy==1.16.4
opencv-python==3.4.2.17
pandas==0.22.0
protobuf==3.6.0
scikit-image==0.14.2
scikit-learn==0.19.1
joblib
dlib==19.9.0
face-recognition==1.2.3
face-recognition-models==0.3.0
Flask==0.12.2
Pillow==5.2.0
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0rc2
ffmpeg-python==0.1.6